+++ dashboard/services/resources.py
@@ -15,6 +15,7 @@
 
 # Service Layer: Process and cache REST resource's responses here.
 
+import logging
 from subprocess import Popen, PIPE
 from collections import OrderedDict
 try:
@@ -33,6 +34,8 @@ from dashboard.services.consume import call_service
 
 __all__ = ['APIResources']
 
+logger = logging.getLogger(__name__)
+
 
 class ResourcesBase(object):
     """Base class for resources"""
@@ -50,9 +53,8 @@ class ResourcesBase(object):
             return api_config['method'](
                 api_config['base_url'], service_resource, *args, **kwargs
             )
-        except (KeyError, Exception):
-            # log error
-            pass
+        except Exception as exp:
+            logger.log(level=10, msg=f"Resource execution caught an exception. Details: {str(exp)}")
         return {}
 
 
@@ -216,6 +218,12 @@ class TransplatformResources(ResourcesBase):
                     pick_releases.append(release['fields']['name'])
                     first_stable_pk = None
         return {'releases': pick_releases}
+
+    @staticmethod
+    @call_service(TRANSPLATFORM_ENGINES[1])
+    def _create_transifex_project(base_url, resource, *url_params, **kwargs):
+        response = kwargs.get('rest_response', {})
+        return response
 
     @staticmethod
     @call_service(TRANSPLATFORM_ENGINES[0])
@@ -395,7 +403,7 @@ class TransplatformResources(ResourcesBase):
     def __push_response(service_resp):
         if service_resp.get('json_content'):
             return True, service_resp['json_content']
-        elif service_resp.get('err_content') or service_resp.get('text'):
+        if service_resp.get('err_content') or service_resp.get('text'):
             return False, service_resp.get('err_content') or \
                 {service_resp.get('status_code', 'Error'): service_resp.get('text')}
         return False, {}
@@ -537,6 +545,25 @@ class TransplatformResources(ResourcesBase):
                 'resources': ['project_details', 'project_jobs'],
                 'combine_results': True,
                 'project': args[0],
+            }
+        }
+        selected_config = method_mapper[translation_platform]
+        return self._execute_method(selected_config, *args, **kwargs)
+
+    def create_project(self, translation_platform, instance_url, *args, **kwargs):
+        """
+        Create a Project at Translation or CI Platform
+        :param translation_platform: Translation Platform API
+        :param instance_url: Translation Platform Server URL
+        :param args: URL Params: list
+        :param kwargs: Keyword Args: dict
+        :return: dict
+        """
+        method_mapper = {
+            TRANSPLATFORM_ENGINES[1]: {
+                'method': self._create_transifex_project,
+                'base_url': instance_url,
+                'resources': ['create_project'],
             }
         }
         selected_config = method_mapper[translation_platform]
@@ -725,6 +752,8 @@ class KojiResources(object):
                 tag_starts_with = 'rhevm'
             elif product.product_slug == RELSTREAM_SLUGS[3]:
                 tag_starts_with = 'satellite'
+            elif product.product_slug == RELSTREAM_SLUGS[4]:
+                tag_starts_with = 'insights'
         elif product.product_slug == RELSTREAM_SLUGS[1]:
             tag_starts_with = 'f'
 
@@ -785,9 +814,9 @@ class KojiResources(object):
         if build and not srpm and not task:
             path = koji.pathinfo.build(build)
             return path[0] if isinstance(path, list) and len(path) > 0 else path
-        elif srpm and not build and not task:
+        if srpm and not build and not task:
             return koji.pathinfo.rpm(srpm)
-        elif task and not build and not srpm:
+        if task and not build and not srpm:
             return koji.pathinfo.task(task)