+++ transtats/middleware.py
@@ -34,7 +34,7 @@ class CustomMiddleware(object):
         """Called just before Django calls the view."""
         request.tenant = settings.TS_AUTH_SYSTEM \
             if settings.TS_AUTH_SYSTEM in RELSTREAM_SLUGS and \
-            view_func.__module__ in ['dashboard.views'] else "default"
+            view_func.__module__ in ['dashboard.views'] else None
 
         return None